replaced all pow calls with powf

+++ platforms/cuda/src/kernels/steinhardt.cu
@@ -23,10 +23,10 @@ extern "C" __global__ void computeSteinhardt(int numParticles, const real4* __re
           real3 positionj=trimTo3(posq[particles[j]]);
           real3 rij= make_real3(positioni.x-positionj.x, positioni.y-positionj.y, positioni.z-positionj.z);
           APPLY_PERIODIC_TO_DELTA(rij);
-          real rij_norm=pow(rij.x*rij.x + rij.y*rij.y + rij.z*rij.z,0.5);
+          real rij_norm=powf(rij.x*rij.x + rij.y*rij.y + rij.z*rij.z,0.5);
 	  if(rij_norm<1.6*CUTOFF){
-	  real rij_pow6=pow((rij_norm-CUTOFF)/1,6);
-          real switch_ij=(1-rij_pow6)/(1-pow(rij_pow6,2));
+	  real rij_pow6=powf((rij_norm-CUTOFF)/1,6);
+          real switch_ij=(1-rij_pow6)/(1-powf(rij_pow6,2));
 	  
           sumN += switch_ij;
           for(int k=0; k<numParticles; k++){
@@ -34,12 +34,12 @@ extern "C" __global__ void computeSteinhardt(int numParticles, const real4* __re
               real3 positionk=trimTo3(posq[particles[k]]);
               real3 rik= make_real3(positioni.x-positionk.x, positioni.y-positionk.y, positioni.z-positionk.z);
               APPLY_PERIODIC_TO_DELTA(rik);
-              real rik_norm=pow(rik.x*rik.x + rik.y*rik.y + rik.z*rik.z,0.5);
+              real rik_norm=powf(rik.x*rik.x + rik.y*rik.y + rik.z*rik.z,0.5);
 	      if(rik_norm<1.6*CUTOFF){
-	      real rik_pow6=pow((rik_norm-CUTOFF)/1,6);
-              real switch_ik=(1-rik_pow6)/(1-pow(rik_pow6,2));
+	      real rik_pow6=powf((rik_norm-CUTOFF)/1,6);
+              real switch_ik=(1-rik_pow6)/(1-powf(rik_pow6,2));
               real rdot = (rij.x*rik.x + rij.y*rik.y + rij.z*rik.z)/(rik_norm*rij_norm);
-              real P6=(231*pow(rdot,6.0)-315*pow(rdot,4.0)+105*pow(rdot,2.0)-5)/16;
+              real P6=(231*powf(rdot,6.0)-315*powf(rdot,4.0)+105*powf(rdot,2.0)-5)/16;
               //M[i] += P6*switch_ik*switch_ij;
               sumM += P6*switch_ik*switch_ij;
             }
@@ -64,28 +64,28 @@ extern "C" __global__ void computeSteinhardtForces(int numParticles, const real4
                  real4 periodicBoxVecX, real4 periodicBoxVecY, real4 periodicBoxVecZ,real* __restrict__ M, real* __restrict__ N, real* __restrict__ F, real Q_tot) {
 
     unsigned int i = blockIdx.x*blockDim.x+threadIdx.x;
-    real prefactor=pow(4*3.14159/13,0.5)/numParticles;
+    real prefactor=powf(4*3.14159/13,0.5)/numParticles;
     while(i < numParticles){
     	//printf("%d\n", i);
 	//printf("Hello from block %d, thread %d\n", blockIdx.x, threadIdx.x);
         real3 positioni=trimTo3(posq[particles[i]]);
 	
 	real3 Ficomp=make_real3(0);
-	real M_prefactor=-1/(2*N[i]*pow(M[i],0.5));
+	real M_prefactor=-1/(2*N[i]*powf(M[i],0.5));
 
         for(int j=0; j<numParticles; j++){
             if( j!=i ){
                 real3 positionj=trimTo3(posq[particles[j]]);
                 real3 rij= make_real3(positioni.x-positionj.x, positioni.y-positionj.y, positioni.z-positionj.z);
                 APPLY_PERIODIC_TO_DELTA(rij);
-                real rij_norm=pow(rij.x*rij.x + rij.y*rij.y + rij.z*rij.z,0.5);
+                real rij_norm=powf(rij.x*rij.x + rij.y*rij.y + rij.z*rij.z,0.5);
 		if(rij_norm < 2.0*CUTOFF){
                 real3 delta_rij_norm=rij/(2*rij_norm);
-		real rij_pow6=pow((rij_norm-CUTOFF)/1,6);
+		real rij_pow6=powf((rij_norm-CUTOFF)/1,6);
                 real switch_ij_numerator=(1-rij_pow6);
-                real switch_ij_denominator=(1-pow(rij_pow6,2));
+                real switch_ij_denominator=(1-powf(rij_pow6,2));
                 real switch_ij=switch_ij_numerator/switch_ij_denominator;
-		real rij_pow5=pow((rij_norm-CUTOFF)/1,5);
+		real rij_pow5=powf((rij_norm-CUTOFF)/1,5);
                 real delta_switch_ij=(6*rij_pow5*switch_ij_denominator-12*rij_pow5*rij_pow6*switch_ij_numerator)/(switch_ij_denominator*switch_ij_denominator);
 		real3 Fjcomp=make_real3(0);
                 for(int k=0; k<numParticles; k++){
@@ -94,19 +94,19 @@ extern "C" __global__ void computeSteinhardtForces(int numParticles, const real4
                         real3 rik= make_real3(positioni.x-positionk.x, positioni.y-positionk.y, positioni.z-positionk.z);
                         APPLY_PERIODIC_TO_DELTA(rik);
 			
-                        real rik_norm=pow(rik.x*rik.x + rik.y*rik.y + rik.z*rik.z,0.5);
+                        real rik_norm=powf(rik.x*rik.x + rik.y*rik.y + rik.z*rik.z,0.5);
 			if(rik_norm < 2.0*CUTOFF){
                         real3 delta_rik_norm=rik/(2*rik_norm);
-			real rik_pow6=pow((rik_norm-CUTOFF)/1,6);
+			real rik_pow6=powf((rik_norm-CUTOFF)/1,6);
                         real switch_ik_numerator=(1-rik_pow6);
-                        real switch_ik_denominator=(1-pow(rik_pow6,2));
+                        real switch_ik_denominator=(1-powf(rik_pow6,2));
                         real switch_ik=switch_ik_numerator/switch_ik_denominator;
-			real rik_pow5=pow((rik_norm-CUTOFF)/1,5);
+			real rik_pow5=powf((rik_norm-CUTOFF)/1,5);
                         real delta_switch_ik=(6*rik_pow5*switch_ik_denominator-12*rik_pow5*rik_pow6*switch_ik_numerator)/(switch_ik_denominator*switch_ik_denominator);
 			
                         real rdot = rij.x*rik.x + rij.y*rik.y + rij.z*rik.z/(rij_norm*rik_norm);
-                        real P6=(231*pow(rdot,6.0)-315*pow(rdot,4.0)+105*pow(rdot,2.0)-5)/16;
-                        real delta_P6=(1386*pow(rdot,5.0)-1260*pow(rdot,3)+210*rdot)/16;
+                        real P6=(231*powf(rdot,6.0)-315*powf(rdot,4.0)+105*powf(rdot,2.0)-5)/16;
+                        real delta_P6=(1386*powf(rdot,5.0)-1260*powf(rdot,3)+210*rdot)/16;
 			real3 delta_rijj= -2*delta_rij_norm + make_real3(2*delta_rij_norm.x*delta_rij_norm.x*delta_rik_norm.x,2*delta_rij_norm.y*delta_rij_norm.y*delta_rik_norm.y,2*delta_rij_norm.z*delta_rij_norm.z*delta_rik_norm.z)/rij_norm;
 			real3 delta_rikk= -2*delta_rik_norm + make_real3(2*delta_rik_norm.x*delta_rik_norm.x*delta_rij_norm.x,2*delta_rik_norm.y*delta_rik_norm.y*delta_rij_norm.y,2*delta_rik_norm.z*delta_rik_norm.z*delta_rij_norm.z)/rik_norm;
 			real3 delta_rijik = -delta_rijj-delta_rikk;
